pyproject.toml
[tool.poetry]
name = "sensor-data-generator"
version = "0.1.0"
description = "A data generator for sensor data in TimescaleDB"
authors = ["baiogit <baioni.federico@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
generate-data = "generator.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.41"
pandas = "^2.0.0"
numpy = "^1.24.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
sqlalchemy-utils = "^0.41.0"
flask = "^3.1.1"
scikit-learn = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pylint = "^3.0.0"
black = "^23.0.0"
isort = "^5.12.0"
